refactor(integration/notion): Update properties payload description and schema import
